Rename query builder file
