+++ lib/postgresql-store.js
@@ -5,7 +5,7 @@ var _ = require('lodash')
 var Pg = require('pg')
 var Uuid = require('node-uuid')
 var RelationalStore = require('./relational-util')
-var QueryBuilder = require('./query_builder')
+var QueryBuilder = require('./query-builder')
 
 var name = 'postgresql-store'
