+++ lib/query-builder.js
